requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
pandas
tensorflow
scikit-learn
google-cloud-storage
pydantic
python-dotenv